requirements.txt
Django
cryptography
Pillow
python-dotenv
PyMySQL
djangorestframework
docx2txt
pypdf
seleniumbase
drf-yasg
django-cors-headers
langchain-community
langchain-chroma
selenium
mysqlclient
webdriver-manager
django-filter
openai
fake-useragent
pandas
langchain_openai

